+++ data_structures.h
@@ -1,17 +1,17 @@
 #include <vector>
 #include <iostream>
 #include <cassert>
-//typedef size_t uint16_t;
+//typedef size_t uint64_t16_t;
 
 class node {
 public:
-	node(int id) : id_(id), num_child_(0) {
+	node(int64_t id) : id_(id), num_child_(0) {
 		//set_null_children();
 		id_ = id;
 		num_child_ = 0;
 	}
 
-	node(int id, size_t num_child, bool is_optical_switch) : id_(id), num_child_(num_child) {
+	node(int64_t id, size_t num_child, bool is_optical_switch) : id_(id), num_child_(num_child) {
 		children_.reserve(num_child_);
 		children_.resize(num_child_);
 		id_ = id;
@@ -22,13 +22,13 @@ public:
 
 	~node() {};
 
-	int get_num_child() const {
-		int return_val = num_child_;
+	int64_t get_num_child() const {
+		int64_t return_val = num_child_;
 		assert(num_child_==children_.size());
 		return return_val;
 	};
 
-	void set_num_child(int num_child) {
+	void set_num_child(int64_t num_child) {
 		if (num_child > 0) {
 			num_child_ = num_child;
 			children_.reserve(num_child_);
@@ -36,7 +36,7 @@ public:
 		}
 	};
 
-	void set_child(int index, node* child_node) {
+	void set_child(int64_t index, node* child_node) {
  		assert(index >= 0 && index < num_child_);
  		//assert(child_node);
 		if (index < num_child_ && index >= 0) {
@@ -44,7 +44,7 @@ public:
 		}
 	};
 
-	node* get_child(int index) const {
+	node* get_child(int64_t index) const {
 
 		node* child = nullptr;
 		assert(index < num_child_ && index >= 0);
@@ -55,7 +55,7 @@ public:
 		//return child;
 	};
 
-	int get_id() const {
+	int64_t get_id() const {
 		return id_;
 	};
 
@@ -68,15 +68,15 @@ private:
 		if (num_child_ == 0) {
 			return;
 		}
-		for (int i = 0; i < num_child_; i++) {
+		for (int64_t i = 0; i < num_child_; i++) {
 			children_[i] = nullptr;
 		}
 	};
 
 private:
-	int num_child_;
+	int64_t num_child_;
 	
-	int id_;
+	int64_t id_;
 	
 	std::vector<node*> children_;
